# docs/swagger.yaml
basePath: /
definitions:
  models.Micropost:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Your API Description
  title: Your API Title
  version: "1.0"
paths:
  /microposts:
    get:
      description: Get all microposts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Micropost'
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List microposts
      tags:
      - microposts
    post:
      consumes:
      - application/json
      description: Create a new micropost
      parameters:
      - description: Micropost object
        in: body
        name: micropost
        required: true
        schema:
          $ref: '#/definitions/models.Micropost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Micropost'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create micropost
      tags:
      - microposts
  /microposts/{id}:
    delete:
      description: Delete a micropost by its ID
      parameters:
      - description: Micropost ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete micropost
      tags:
      - microposts
    get:
      description: Get a micropost by its ID
      parameters:
      - description: Micropost ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Micropost'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get micropost by ID
      tags:
      - microposts
    put:
      consumes:
      - application/json
      description: Update a micropost by its ID
      parameters:
      - description: Micropost ID
        in: path
        name: id
        required: true
        type: string
      - description: Micropost object
        in: body
        name: micropost
        required: true
        schema:
          $ref: '#/definitions/models.Micropost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Micropost'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update micropost
      tags:
      - microposts
swagger: "2.0"